+++ lib/export/coldp/files/name.rb
@@ -1,9 +1,5 @@
-# See
-# * https://api.catalogue.life/vocab/rank
-# * https://github.com/SpeciesFileGroup/taxonworks/issues/1040
 module Export::Coldp::Files::Name
 
-  # We don't cover ICNCP?
   def self.code_field(taxon_name)
     case taxon_name.nomenclatural_code
     when :iczn
@@ -13,7 +9,7 @@ module Export::Coldp::Files::Name
     when :icnp
       'ICNP'
     when :icvcn
-      'ICVCN' # doublecheck
+      'ICVCN'
     end
   end
 
@@ -21,44 +17,40 @@ module Export::Coldp::Files::Name
     Utilities::Strings.nil_squish_strip(taxon_name.notes.collect{|n| n.text}.join('; ')) # remarks - !! check for tabs
   end
 
-  def self.remarks_field(taxon_name)
-    Utilities::Strings.nil_squish_strip(taxon_name.notes.collect{|n| n.text}.join('; ')) # remarks - !! check for tabs
-  end
-
-  # Not used in main loop, for reference only
-  def published_in_id(taxon_name)
-    taxon_name.source&.id
-  end
-
-  # @return Boolean
-  #    true if original combination of the protonym
-
+  # @return String
   def self.authorship_field(taxon_name, original)
     original ? taxon_name.original_author_year : taxon_name.cached_author_year
   end
 
-  # @return [String, nil]
   # https://api.catalogue.life/vocab/nomStatus
-  # TODO, move this to the concern
-  #   https://github.com/SpeciesFileGroup/taxonworks/issues/1040
+  # @return [String, nil]
+  # @params taxon_name [TaxonName]
+  #   any TaxonName 
   def self.nom_status_field(taxon_name)
     case taxon_name.type
     when 'Combination'
       'chresonym'
     else
-      c = taxon_name.taxon_name_classifications_for_statuses.order_by_youngest_source_first.first
-      c ? c.class::NOMEN_URI : nil
+      if taxon_name.is_valid?
+        ::TaxonName::NOMEN_VALID[taxon_name.nomenclatural_code]
+      else
+        c = taxon_name.taxon_name_classifications_for_statuses.order_by_youngest_source_first.first
+        c ? c.class::NOMEN_URI : nil # We should also infer status from TaxonNameRelationship see 
+      end
     end
   end
 
-  # Is never a Combination, handles original combination
-  def self.add_current_combination(t, csv)
+  # Invalid Protonyms are rendered only as their original Combination 
+  # @param t [An invalid Protonym]
+  def self.add_original_combination(t, csv)
     e = t.original_combination_elements
     csv << [
-      ::Export::Coldp.current_taxon_name_id(t),                               # ID
+      ::Export::Coldp.reified_id(t),                                          # ID
+      t.id,                                                                   # basionymID, always nil, this is the original
       t.cached_original_combination,                                          # scientificName
-      t.cached_author_year,                                                   # authorship
+      authorship_field(t, true),                                              # authorship
       t.rank,                                                                 # rank
+      nil,                                                                    # uninomial
       (e[:genus] =~ /NOT SPECIFIED/) ? nil : e[:genus]&.join(' '),            # genus
       (e[:subgenus] =~ /NOT SPECIFIED/) ? nil : e[:subgenus]&.join(' '),      # subgenus
       (e[:species] =~ /NOT SPECIFIED/) ? nil : e[:species]&.join(' '),        # species
@@ -66,7 +58,7 @@ module Export::Coldp::Files::Name
       nil,                                                                    # publishedInID   |
       nil,                                                                    # publishedInPage |-- Decisions is that these add to Synonym table
       nil,                                                                    # publishedInYear |
-      false,                                                                  # original
+      true,                                                                   # original
       code_field(t),                                                          # code
       nil,                                                                    # status https://api.catalogue.life/vocab/nomStatus
       nil,                                                                    # link (probably TW public or API)
@@ -74,14 +66,17 @@ module Export::Coldp::Files::Name
     ]
   end
 
+  # @params otu [Otu]
+  #   the top level OTU 
   def self.generate(otu, reference_csv = nil)
     CSV.generate(col_sep: "\t") do |csv|
-
       csv << %w{
         ID
+        basionymID
         scientificName
         authorship
         rank
+        uninomial
         genus
         infragenericEpithet
         specificEpithet
@@ -96,33 +91,44 @@ module Export::Coldp::Files::Name
         remarks
       }
 
-      otu.taxon_name.self_and_descendants.each do |t|
-        source = t.source # published_in_id_field
+      otu.taxon_name.self_and_descendants.each do |name|
 
-        original = Export::Coldp.original_field(t)
+        if name.is_valid?
+          data = ::Catalog::Nomenclature::Entry.new(name)
+          data.names.each do |t|
+            source = t.source
 
-        csv << [
-          t.id,                                      # ID
-          t.cached,                                  # scientificName
-          authorship_field(t, original),             # authorship
-          t.rank,                                    # rank
-          t.ancestor_at_rank('genus', true)&.name,         # genus
-          t.ancestor_at_rank('subgenus', true)&.name,      # infragenericEpithet
-          t.ancestor_at_rank('species', true)&.name,       # specificEpithet
-          t.ancestor_at_rank('subspecies', true)&.name,    # infraspecificEpithet
-          source&.id,                                # publishedInID
-          source&.pages,                             # publishedInPage
-          t.year_of_publication,                     # publishedInYear
-          original,                                  # original
-          code_field(t),                             # code
-          nom_status_field(t),                       # nomStatus
-          nil,                                       # link (probably TW public or API)
-          remarks_field(t),                          # remarks
-        ]
+            original = Export::Coldp.original_field(t) # no parens
+            higher = !(t.type == 'Combination') && !t.is_species_rank?
 
-        add_current_combination(t, csv) if !original && !t.cached_original_combination.blank?
+            if higher || (t.is_valid? || t.type == 'Combination')
+              csv << [
+                t.id,                                                          # ID
+                (original ? nil : ::Export::Coldp.basionym_id(t)),             # basionymID
+                t.cached,                                                      # scientificName
+                t.cached_author_year,                                          # authorship
+                t.rank,                                                        # rank
+                (higher ? t.cached : nil),                                     # uninomial
+                (higher ? nil : t.ancestor_at_rank('genus', true)&.name),      # genus and below - IIF species or lower
+                (higher ? nil : t.ancestor_at_rank('subgenus', true)&.name),   # infragenericEpithet
+                (higher ? nil : t.ancestor_at_rank('species', true)&.name),    # specificEpithet
+                (higher ? nil : t.ancestor_at_rank('subspecies', true)&.name), # infraspecificEpithet
+                source&.id,                                                    # publishedInID
+                source&.pages,                                                 # publishedInPage
+                t.year_of_publication,                                         # publishedInYear
+                original,                                                      # original
+                code_field(t),                                                 # code
+                nom_status_field(t),                                           # nomStatus
+                nil,                                                           # link (probably TW public or API)
+                remarks_field(t),                                              # remarks
+              ]
 
-        Export::Coldp::Files::Reference.add_reference_rows([source].compact, reference_csv) if reference_csv
+              Export::Coldp::Files::Reference.add_reference_rows([source].compact, reference_csv) if reference_csv && source
+            else
+              add_original_combination(t, csv)
+            end
+          end
+        end
       end
     end
   end